requirements.txt
requests>=2.28.0
pyyaml>=6.0
python-dotenv>=0.19.0
jinja2>=3.0.0
pytest>=7.0.0
pytest-mock>=3.10.0
celery>=5.2.0
redis>=4.3.0